requirements.txt
# Core dependencies
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.24.0

# LLM API access
openai>=1.3.0
anthropic>=0.8.0

# Vector database and embeddings
chromadb>=0.4.0

# Natural Language Processing
spacy>=3.7.0
# Install spaCy model with:
# python -m spacy download en_core_web_sm

# Document processing
python-docx>=0.8.10

# Visualization and UI components
matplotlib>=3.7.0
plotly>=5.15.0

# Utility libraries
tqdm>=4.65.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development tools
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0